add used slot count for reference and item array
mpt++
  next position allways empty after first is found

+++ mpt++/graphic.h
@@ -146,7 +146,7 @@ public:
     // layout (de)registration
     virtual int addLayout(Layout *, bool = true);
     virtual int removeLayout(const Layout *);
-    int layoutCount() const;
+    long layoutCount() const;
 
     // create new layout
     virtual Layout *createLayout();
